Simplify instantiation and hooking into flask response handling
In #8 @soxofaan kindly raised the attention that the way this module is
hooking into flask's `after_this_request()` handling is unnecessarily
complex.

Simplified this drastically. Thanks a lot @soxofaan!

Caveat: OnlyIfEvaluatorBase interface changed.

Fixes #8
