+++ src/flask_cachecontrol/callback.py
@@ -7,17 +7,27 @@
     :license: BSD, see LICENSE for more details.
 """
 
+from abc import ABCMeta, abstractmethod
 from datetime import datetime
 
-from .after_this_request import CallbackBase
 from .error import CacheControlAttributeInvalidError
+
+
+class CallbackBase(metaclass=ABCMeta):
+    def __call__(self, response):
+        self._process_response(response)
+        return response
+
+    @abstractmethod
+    def _process_response(self, response):
+        pass
 
 
 class SetCacheControlHeadersFromTimedeltaCallback(CallbackBase):
     def __init__(self, timedelta):
         self._timedelta = timedelta
 
-    def __call__(self, response):
+    def _process_response(self, response):
         response.expires = datetime.utcnow() + self._timedelta
         response.cache_control.max_age = int(self._timedelta.total_seconds())
 
@@ -26,7 +36,7 @@ class SetCacheControlHeadersCallback(CallbackBase):
     def __init__(self, **cache_control_kw):
         self._cache_control_kw = cache_control_kw
 
-    def __call__(self, response):
+    def _process_response(self, response):
         cache_control = response.cache_control
         for attr_name, value in self._cache_control_kw.items():
             if not hasattr(cache_control, attr_name):
@@ -35,7 +45,7 @@ class SetCacheControlHeadersCallback(CallbackBase):
 
 
 class SetCacheControlHeadersForNoCachingCallback(CallbackBase):
-    def __call__(self, response):
+    def _process_response(self, response):
         response.cache_control.max_age = 0
         response.cache_control.no_cache = True
         response.cache_control.no_store = True
@@ -48,5 +58,5 @@ class SetVaryHeaderCallback(CallbackBase):
         # generate string version once for object lifetime
         self._vary_str = ','.join(vary)
 
-    def __call__(self, response):
+    def _process_response(self, response):
         response.vary = self._vary_str
